make sure node type method

+++ python/tensorstack/frontend/onnx.py
@@ -44,8 +44,8 @@ def pooling2d_padding(name, x, padding, ksize, stride, auto_pad=Name.NOTSET):
     padding = zoo.to_const(padding, "padding")
 
     # input
-    ksize = zoo.to_node(ksize, name="_const_" + name + "_ksize", device=device.CPU)
-    stride = zoo.to_node(stride, name="_const_" + name + "_stride", device=device.CPU)
+    ksize = zoo.to_node(ksize, name="_const_" + name + "_ksize", dtype=numpy.int32, device=device.CPU)
+    stride = zoo.to_node(stride, name="_const_" + name + "_stride", dtype=numpy.int32, device=device.CPU)
 
     # operator
     node = menu.op(name=name, op_name=Name.Layer.pooling2d_padding, inputs=[x, ksize, stride])
@@ -70,8 +70,8 @@ def pooling2d(name, x, ksize, stride, type=zoo.Type.pooling_type.max, format=zoo
     static_padding = zoo.to_const(padding, "padding")
 
     # input
-    ksize = zoo.to_node(ksize, name="_const_" + name + "_ksize", device=device.CPU)
-    stride = zoo.to_node(stride, name="_const_" + name + "_stride", device=device.CPU)
+    ksize = zoo.to_node(ksize, name="_const_" + name + "_ksize", dtype=numpy.int32, device=device.CPU)
+    stride = zoo.to_node(stride, name="_const_" + name + "_stride", dtype=numpy.int32, device=device.CPU)
 
     # operator
     dynamic_padding = pooling2d_padding(name="_op_" + name + "_onnx_padding",
@@ -84,7 +84,7 @@ def pooling2d(name, x, ksize, stride, type=zoo.Type.pooling_type.max, format=zoo
 def gather(name, x, indices, axis=0):
     assert isinstance(x, Node)
 
-    indices = zoo.to_node(indices, name="_const_" + name + "_indices", device=device.CPU)
+    indices = zoo.to_node(indices, name="_const_" + name + "_indices", dtype=numpy.int32, device=device.CPU)
     axis = zoo.to_const(axis, "axis")
 
     # operator
