+++ python/tensorstack/frontend/tf.py
@@ -43,8 +43,8 @@ def pooling2d_padding(name, x, padding, ksize, stride, format=zoo.Name.NCHW, pad
     padding = zoo.to_const(padding, "padding")
 
     # input
-    ksize = zoo.to_node(ksize, name="_const_" + name + "_ksize", device=device.CPU)
-    stride = zoo.to_node(stride, name="_const_" + name + "_stride", device=device.CPU)
+    ksize = zoo.to_node(ksize, name="_const_" + name + "_ksize", dtype=numpy.int32, device=device.CPU)
+    stride = zoo.to_node(stride, name="_const_" + name + "_stride", dtype=numpy.int32, device=device.CPU)
 
     # operator
     node = menu.op(name=name, op_name=Name.Layer.pooling2d_padding, inputs=[x, ksize, stride])
@@ -68,8 +68,8 @@ def pooling2d(name, x, ksize, stride, type=zoo.Type.pooling_type.max, format=zoo
     static_padding = zoo.to_const(padding, "padding")
 
     # input
-    ksize = zoo.to_node(ksize, name="_const_" + name + "_ksize", device=device.CPU)
-    stride = zoo.to_node(stride, name="_const_" + name + "_stride", device=device.CPU)
+    ksize = zoo.to_node(ksize, name="_const_" + name + "_ksize", dtype=numpy.int32, device=device.CPU)
+    stride = zoo.to_node(stride, name="_const_" + name + "_stride", dtype=numpy.int32, device=device.CPU)
 
     # operator
     dynamic_padding = pooling2d_padding(name="_op_" + name + "_tf_padding",
